Move range-picker logic to other directory
